requirements.txt
pyTelegramBotAPI==4.26.0
python-dotenv==1.1.0
yt-dlp==2025.3.31
requests==2.32.3
